# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jengo Building Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /buildings/{buildingId}/units:
    get:
      operationId: listUnits
      tags: [units]
      summary: List units in a building
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create a unit
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUnitBody"
      responses:
        "201":
          description: Created unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{id}:
    get:
      operationId: getUnit
      tags: [units]
      summary: Get a unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      operationId: updateUnit
      tags: [units]
      summary: Update a unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUnitBody"
      responses:
        "200":
          description: Updated unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"

  # ── Residents ──────────────────────────────────────────────────────────────
...
  /visitors:
    get:
      operationId: listVisitors
      tags: [visitors]
      summary: List visitors
      parameters:
        - name: buildingId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, checked_in, checked_out, denied]
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of visitors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visitor"
    post:
      operationId: createVisitor
      tags: [visitors]
      summary: Pre-clear a visitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVisitorBody"
      responses:
        "201":
          description: Created visitor clearance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitor"
  /visitors/{id}:
    get:
      operationId: getVisitor
      tags: [visitors]
      summary: Get visitor details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Visitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitor"
    patch:
      operationId: updateVisitor
      tags: [visitors]
      summary: Update visitor status (check-in, check-out, approve, deny)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVisitorBody"
      responses:
        "200":
          description: Updated visitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitor"
  /visitors/today:
    get:
      operationId: getTodayVisitors
      tags: [visitors]
      summary: Get today's visitor summary for gate security
      parameters:
        - name: buildingId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Today's visitors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayVisitorsSummary"

  # ── Payments ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Unit:
      type: object
      properties:
        id:
          type: integer
        buildingId:
          type: integer
        unitNumber:
          type: string
        floor:
          type: integer
        bedrooms:
          type: integer
        status:
          type: string
          enum: [occupied, vacant, maintenance]
        monthlyRent:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, buildingId, unitNumber, status, createdAt]
    CreateUnitBody:
      type: object
      properties:
        unitNumber:
          type: string
        floor:
          type: integer
        bedrooms:
          type: integer
        status:
          type: string
          enum: [occupied, vacant, maintenance]
        monthlyRent:
          type: number
      required: [unitNumber, status]
    UpdateUnitBody:
      type: object
      properties:
        status:
          type: string
          enum: [occupied, vacant, maintenance]
        monthlyRent:
          type: number

    # ── Resident ──
...
    Visitor:
      type: object
      properties:
        id:
          type: integer
        buildingId:
          type: integer
        unitId:
          type: integer
        residentId:
          type: integer
        visitorName:
          type: string
        visitorPhone:
          type: string
        visitorIdNumber:
          type: string
        purpose:
          type: string
        expectedDate:
          type: string
          format: date
        expectedTime:
          type: string
        status:
          type: string
          enum: [pending, approved, checked_in, checked_out, denied]
        checkInTime:
          type: string
          format: date-time
        checkOutTime:
          type: string
          format: date-time
        securityNote:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, buildingId, visitorName, expectedDate, status, createdAt]
    CreateVisitorBody:
      type: object
      properties:
        buildingId:
          type: integer
        unitId:
          type: integer
        residentId:
          type: integer
        visitorName:
          type: string
        visitorPhone:
          type: string
        visitorIdNumber:
          type: string
        purpose:
          type: string
        expectedDate:
          type: string
          format: date
        expectedTime:
          type: string
      required: [buildingId, visitorName, expectedDate]
    UpdateVisitorBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, checked_in, checked_out, denied]
        securityNote:
          type: string
        checkInTime:
          type: string
          format: date-time
        checkOutTime:
          type: string
          format: date-time
    TodayVisitorsSummary:
      type: object
      properties:
        expected:
          type: integer
        checkedIn:
          type: integer
        checkedOut:
          type: integer
        pending:
          type: integer
        denied:
          type: integer
        visitors:
          type: array
          items:
            $ref: "#/components/schemas/Visitor"
      required: [expected, checkedIn, checkedOut, pending, denied, visitors]

    # ── Payment ──
...